fix: correct debian/ubuntu install instructions (#3312)
* fix: correct debian/ubuntu install instructions

This consolidates the Debian and Ubutnu install instructions into one
section. The install instructions can be the same for both as they use
the same package manager (i.e. apt).

This removes the use of the now deprecated pipe to apt-key.

This removes references to starting the service, as the service starts
on install and does not require an extra step. Also all supported
operating systems are using systemd at this point so no need to call
that out seperatly.

Removes an extra 'sudo' in deb instructions.

* missing word