fix: Clarify valuecounter aggregator
The aggregator only applies to fields and not tags. The current example
use-case explains how to use the aggregator, but the example itself is
suggesting to use the aggreagator on tags that are generated. We got an
issue complaining about this and I'd like to just remove it and instead
clarify fields only.