influxctl authentication note (#5102)
* influxctl authentication note

This just adds a helpful pointer to eliminate an annoying and unnecessary warning message that may trip up parsing JSON output from the utility.  The note is Linux specific.  It may be Ubuntu specific as well, but this is one of the most popular distros.

* Update content/influxdb/cloud-dedicated/reference/cli/influxctl/_index.md

Co-authored-by: Jason Stirnaman <stirnamanj@gmail.com>

* Update content/influxdb/cloud-dedicated/reference/cli/influxctl/_index.md

* Update _index.md

Apply suggestion

Co-authored by: Jamie Strandboge <jdstrand@users.noreply.github.com>

* Update content/influxdb/cloud-dedicated/reference/cli/influxctl/_index.md

---------

Co-authored-by: Jason Stirnaman <stirnamanj@gmail.com>