chore: always verify GPG fingerprint
The influxdata-archive.key is designed to have a good user experience
during key rotations, but in order to do that, we need to properly
verify the key file's fingerprint and not its sha256sum (since the
primary key's fingerprint won't change with key rotations, but the
sha256sum necessarily will).

Update documentation for this to establish best practice and prepare
users for the upcoming key rotation.

Adjust dockerfiles to use this methodology.